move win_utils.h to aux
